nl2sql/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pymysql==1.1.0
tabulate==0.9.0
sshtunnel==0.4.0
paramiko==2.12.0
